Split up DQA test into unit tests and integration tests
